// src/models/user.model.ts
export type User = {
    name: string;
    username: string;
    bolaPercentage: number;
    accountType: string;
    creationDate: string;
    favoriteTeams: string[];
    leagues: string[];
    championships: string[];
    totalDeposited: number;
    email: string;
    location: string;
    avatar?: string;
  }